Need to delte encrypt
so after a long while of trouble shooting I will be removing this  "encryption" thing (copy and pasted it) and just going with another method on how to store user data. I will delete and add that the next commit. Else i just need to add where it shows passwords and thats about it.